Ported from GUIDE to Appdesigner format
Lots of little changes, but also handles are 'handled' differently so its unclear if this is going to end up having unforeseen consequences. GetUIdata and refreshing the gui is gone as app handle covers that now. Testing needed.
Calibration loading and saving changed to be hardware and subject specific, could also add a datastamp on saving